+++ docs/changelog.rst
@@ -6,6 +6,9 @@ Changelog
 `Unreleased`_
 -------------
 
+`2.3.4`_ - 2020-09-11
+---------------------
+
 Changed
 ~~~~~~~
 
@@ -1220,7 +1223,8 @@ Fixed
 
 - Initial public release
 
-.. _Unreleased: https://github.com/schemathesis/schemathesis/compare/v2.3.3...HEAD
+.. _Unreleased: https://github.com/schemathesis/schemathesis/compare/v2.3.4...HEAD
+.. _2.3.4: https://github.com/schemathesis/schemathesis/compare/v2.3.3...v2.3.4
 .. _2.3.3: https://github.com/schemathesis/schemathesis/compare/v2.3.2...v2.3.3
 .. _2.3.2: https://github.com/schemathesis/schemathesis/compare/v2.3.1...v2.3.2
 .. _2.3.1: https://github.com/schemathesis/schemathesis/compare/v2.3.0...v2.3.1